Infer file type

The pxFileTypes() function inferres mass spectrometry and proteomics file types based on a currated table of file types and associated patterns. This table can be accessed with fileTypes(). See the examples below for the content of the tables.

The file types table is generated by scripts/make_fileTypes.R.

fileTypes()

pxFileTypes(fls, types = fileTypes())

Arguments

fls

character() of file names whose types need to be inferred based on their file extenstion.

types

data.frame of file types. Default is fileTypes().

Value

A data.frame with the filenames and their inferred types.

Author

Laurent Gatto with contributions from Dr. Samuel Wein to identify several file types.

Examples

fileTypes()
#>              type                                  ext
#> 1         archive                                  zip
#> 2         archive                               tar.gz
#> 3         archive                                  rar
#> 4         archive                                   7z
#> 5         archive                                  tgz
#> 6         archive                                  ZIP
#> 7         archive                                  tar
#> 8         archive                                  tgz
#> 9         archive                                  RAR
#> 10        archive                                  bz2
#> 11        archive                           webarchive
#> 12        archive                                   gz
#> 13        archive                                   xy
#> 14            doc                                  doc
#> 15            doc                                  pdf
#> 16            doc                                  PDF
#> 17            doc                                  ppt
#> 18            doc                                  odt
#> 19            doc                                 docx
#> 20            doc                                 pptx
#> 21            doc                                  rtf
#> 22            doc                                 html
#> 23            doc                                 html
#> 24            doc                                  htm
#> 25            doc                                shtml
#> 26            doc                               readme
#> 27            doc                                  txt
#> 28            doc                                   md
#> 29         binary                                  raw
#> 30         binary                                  Raw
#> 31         binary                                  RAW
#> 32         binary                                    d
#> 33         binary                                d.zip
#> 34         binary                              raw.zip
#> 35         binary                               raw.gz
#> 36         binary                                 wiff
#> 37         binary                                wiff2
#> 38         binary                            wiff.scan
#> 39         binary                         wiff.1.~idx2
#> 40         binary                             wiff.mtd
#> 41         binary                                  t2d
#> 42            raw                                 mzML
#> 43            raw                              mzML.gz
#> 44            raw                             mzML.zip
#> 45            raw                                mzXML
#> 46            raw                             mzXML.gz
#> 47            raw                            mzXML.zip
#> 48            raw                                mzxml
#> 49            raw                                TraML
#> 50            raw                                traML
#> 51            raw                                traml
#> 52            raw                               netCDF
#> 53            raw                               mzData
#> 54            raw                               mzdata
#> 55            raw                                  mz5
#> 56            raw                                imzML
#> 57            pkl                                  mgf
#> 58            pkl                                  MGF
#> 59            pkl                               mgf.gz
#> 60            pkl                               MGF.gz
#> 61            pkl                                  apl
#> 62            pkl                                  pkl
#> 63            pkl                               pkl.gz
#> 64            pkl                                  PKL
#> 65            fas                                  fas
#> 66            fas                                fasta
#> 67            fas                                   fa
#> 68            fas                                  faa
#> 69            fas                                FASTA
#> 70            fas                                fasts
#> 71            fas                             FALSE.gz
#> 72            fas                           FALSTA.zip
#> 73            fas                             fasta.gz
#> 74            fas                            fasta.zip
#> 75            fas                                fa.gz
#> 76            fas                               fa.zip
#> 77            fas                               faa.gz
#> 78            fas                              faa.zip
#> 79         reflib                                 blib
#> 80         reflib                                 elib
#> 81         reflib                                 dlib
#> 82         reflib                                  msp
#> 83             id                            mzIdentML
#> 84             id                                 mzID
#> 85             id                              mzID.gz
#> 86             id                                 mzid
#> 87             id                              mzid.gz
#> 88             id                             mzid.zip
#> 89             id                                  dat
#> 90             id                               dat.gz
#> 91             id                              dat.zip
#> 92             id                                idXML
#> 93             id                                  omx
#> 94             id                                  sqt
#> 95             id                                 pcml
#> 96             id                                  dta
#> 97             id                              dta.tgz
#> 98             id                          dta.tar.bz2
#> 99            tbl                                  csv
#> 100           tbl                                  tsv
#> 101           tbl                                  xls
#> 102           tbl                                 xlsx
#> 103           tbl                                 XLSX
#> 104           tbl                                 xlsb
#> 105           tbl                                  ssv
#> 106           tbl                               csv.gz
#> 107           tbl                               tsv.gz
#> 108           tbl                               psmtsv
#> 109           tbl                                delim
#> 110           tbl                              tabular
#> 111         mztab                                mztab
#> 112         mztab                             mztab.gz
#> 113         mztab                                mzTab
#> 114         mztab                             mzTab.gz
#> 115         mztab                              mzTabNA
#> 116         mztab                           -mztab.txt
#> 117         mztab                           _mztab.txt
#> 118         mztab                            mztab.txt
#> 119           fig                                  png
#> 120           fig                                  jpg
#> 121           fig                                 jpeg
#> 122           fig                                 tiff
#> 123           fig                                  TIF
#> 124           fig                                  tif
#> 125           fig                                  gif
#> 126           fig                                  PNG
#> 127           fig                                  JPG
#> 128           fig                                  svg
#> 129           xml                                  xml
#> 130           xml                               xml.gz
#> 131       prophet                               pepXML
#> 132       prophet                              protXML
#> 133        bruker                                  yep
#> 134        bruker                                  baf
#> 135      scaffold                                  sf3
#> 136      scaffold                                 sptm
#> 137      scaffold                                 sfdb
#> 138      scaffold                                 sdia
#> 139      scaffold                                metdb
#> 140            pd                             pdResult
#> 141            pd                                  msf
#> 142            pd                         pdResultView
#> 143            pd                              msfView
#> 144            pd                           pdAnalysis
#> 145            pd                       pdProcessingWF
#> 146            pd                        pdConsensusWF
#> 147            pd                              pdStudy
#> 148            pd                          pdStudy.bak
#> 149       sequest                                  ms1
#> 150       sequest                                  ms2
#> 151       sequest                                  srf
#> 152       sequest                                  out
#> 153       sequest                              out.tgz
#> 154       sequest                          out.tar.bz2
#> 155  proteinpilot                                group
#> 156    progenesis              ProgenesisQIPExperiment
#> 157    progenesis                 ProgenesisQIPArchive
#> 158    progenesis             ProgenesisLcmsExperiment
#> 159    progenesis ProgenesisQIPMultiFractionExperiment
#> 160       skyline                                  sky
#> 161       skyline                             sky.view
#> 162       skyline                                 skyd
#> 163       skyline                                 skyl
#> 164   spectronaut                                  sne
#> 165   spectronaut                                htrms
#> 166 peptideshaker                                 cpsx
#> 167        params                               params
#> 168        params                                 apar
#> 169        params                                 knwf
#> 170        params                                 json
#> 171          code                                    R
#> 172          code                                   py
#> 173          code                                    r
#> 174          code                                   pl
#> 175          code                                   js
#> 176          code                                  jar
#> 177          code                                  Rmd
#> 178          code                                   sh
#> 179          code                                ipynb
#> 180          data                                RData
#> 181          data                                  RDS
#> 182          data                               sqlite
#> 183          data                                   h5
#> 184           chk                                  md5
#> 185           chk                                cksum
#> 186           chk                               chksum
#> 187           gen                                  gtf
#> 188           gen                                  gff
#> 189           gen                                fastq
#> 190           gen                                  vcf
#> 191           gen                                plink
#> 192           gen                                  res
#>                                      pattern
#> 1                                    \\.zip$
#> 2                               \\.tar\\.gz$
#> 3                                    \\.rar$
#> 4                                     \\.7z$
#> 5                                    \\.tgz$
#> 6                                    \\.ZIP$
#> 7                                    \\.tar$
#> 8                                    \\.tgz$
#> 9                                    \\.RAR$
#> 10                                   \\.bz2$
#> 11                            \\.webarchive$
#> 12                                    \\.gz$
#> 13                                    \\.xy$
#> 14                                   \\.doc$
#> 15                                   \\.pdf$
#> 16                                   \\.PDF$
#> 17                                   \\.ppt$
#> 18                                   \\.odt$
#> 19                                  \\.docx$
#> 20                                  \\.pptx$
#> 21                                   \\.rtf$
#> 22                                  \\.html$
#> 23                                  \\.html$
#> 24                                   \\.htm$
#> 25                                 \\.shtml$
#> 26                                \\.readme$
#> 27                                   \\.txt$
#> 28                                    \\.md$
#> 29                                   \\.raw$
#> 30                                   \\.Raw$
#> 31                                   \\.RAW$
#> 32                                     \\.d$
#> 33                               \\.d\\.zip$
#> 34                             \\.raw\\.zip$
#> 35                              \\.raw\\.gz$
#> 36                                  \\.wiff$
#> 37                                 \\.wiff2$
#> 38                           \\.wiff\\.scan$
#> 39                      \\.wiff\\.1\\.~idx2$
#> 40                            \\.wiff\\.mtd$
#> 41                                   \\.t2d$
#> 42                                  \\.mzML$
#> 43                             \\.mzML\\.gz$
#> 44                            \\.mzML\\.zip$
#> 45                                 \\.mzXML$
#> 46                            \\.mzXML\\.gz$
#> 47                           \\.mzXML\\.zip$
#> 48                                 \\.mzxml$
#> 49                                 \\.TraML$
#> 50                                 \\.traML$
#> 51                                 \\.traml$
#> 52                                \\.netCDF$
#> 53                                \\.mzData$
#> 54                                \\.mzdata$
#> 55                                   \\.mz5$
#> 56                                 \\.imzML$
#> 57                                   \\.mgf$
#> 58                                   \\.MGF$
#> 59                              \\.mgf\\.gz$
#> 60                              \\.MGF\\.gz$
#> 61                                   \\.apl$
#> 62                                   \\.pkl$
#> 63                              \\.pkl\\.gz$
#> 64                                   \\.PKL$
#> 65                                   \\.fas$
#> 66                                 \\.fasta$
#> 67                                    \\.fa$
#> 68                                   \\.faa$
#> 69                                 \\.FASTA$
#> 70                                 \\.fasts$
#> 71                            \\.FALSE\\.gz$
#> 72                          \\.FALSTA\\.zip$
#> 73                            \\.fasta\\.gz$
#> 74                           \\.fasta\\.zip$
#> 75                               \\.fa\\.gz$
#> 76                              \\.fa\\.zip$
#> 77                              \\.faa\\.gz$
#> 78                             \\.faa\\.zip$
#> 79                                  \\.blib$
#> 80                                  \\.elib$
#> 81                                  \\.dlib$
#> 82                                   \\.msp$
#> 83                             \\.mzIdentML$
#> 84                                  \\.mzID$
#> 85                             \\.mzID\\.gz$
#> 86                                  \\.mzid$
#> 87                             \\.mzid\\.gz$
#> 88                            \\.mzid\\.zip$
#> 89                                   \\.dat$
#> 90                              \\.dat\\.gz$
#> 91                             \\.dat\\.zip$
#> 92                                 \\.idXML$
#> 93                                   \\.omx$
#> 94                                   \\.sqt$
#> 95                                  \\.pcml$
#> 96                                   \\.dta$
#> 97                             \\.dta\\.tgz$
#> 98                       \\.dta\\.tar\\.bz2$
#> 99                                   \\.csv$
#> 100                                  \\.tsv$
#> 101                                  \\.xls$
#> 102                                 \\.xlsx$
#> 103                                 \\.XLSX$
#> 104                                 \\.xlsb$
#> 105                                  \\.ssv$
#> 106                             \\.csv\\.gz$
#> 107                             \\.tsv\\.gz$
#> 108                               \\.psmtsv$
#> 109                                \\.delim$
#> 110                              \\.tabular$
#> 111                                \\.mztab$
#> 112                           \\.mztab\\.gz$
#> 113                                \\.mzTab$
#> 114                           \\.mzTab\\.gz$
#> 115                              \\.mzTabNA$
#> 116                            -mztab\\.txt$
#> 117                            _mztab\\.txt$
#> 118                          \\.mztab\\.txt$
#> 119                                  \\.png$
#> 120                                  \\.jpg$
#> 121                                 \\.jpeg$
#> 122                                 \\.tiff$
#> 123                                  \\.TIF$
#> 124                                  \\.tif$
#> 125                                  \\.gif$
#> 126                                  \\.PNG$
#> 127                                  \\.JPG$
#> 128                                  \\.svg$
#> 129                                  \\.xml$
#> 130                             \\.xml\\.gz$
#> 131                               \\.pepXML$
#> 132                              \\.protXML$
#> 133                                  \\.yep$
#> 134                                  \\.baf$
#> 135                                  \\.sf3$
#> 136                                 \\.sptm$
#> 137                                 \\.sfdb$
#> 138                                 \\.sdia$
#> 139                                \\.metdb$
#> 140                             \\.pdResult$
#> 141                                  \\.msf$
#> 142                         \\.pdResultView$
#> 143                              \\.msfView$
#> 144                           \\.pdAnalysis$
#> 145                       \\.pdProcessingWF$
#> 146                        \\.pdConsensusWF$
#> 147                              \\.pdStudy$
#> 148                        \\.pdStudy\\.bak$
#> 149                                  \\.ms1$
#> 150                                  \\.ms2$
#> 151                                  \\.srf$
#> 152                                  \\.out$
#> 153                            \\.out\\.tgz$
#> 154                      \\.out\\.tar\\.bz2$
#> 155                                \\.group$
#> 156              \\.ProgenesisQIPExperiment$
#> 157                 \\.ProgenesisQIPArchive$
#> 158             \\.ProgenesisLcmsExperiment$
#> 159 \\.ProgenesisQIPMultiFractionExperiment$
#> 160                                  \\.sky$
#> 161                           \\.sky\\.view$
#> 162                                 \\.skyd$
#> 163                                 \\.skyl$
#> 164                                  \\.sne$
#> 165                                \\.htrms$
#> 166                                 \\.cpsx$
#> 167                               \\.params$
#> 168                                 \\.apar$
#> 169                                 \\.knwf$
#> 170                                 \\.json$
#> 171                                    \\.R$
#> 172                                   \\.py$
#> 173                                    \\.r$
#> 174                                   \\.pl$
#> 175                                   \\.js$
#> 176                                  \\.jar$
#> 177                                  \\.Rmd$
#> 178                                   \\.sh$
#> 179                                \\.ipynb$
#> 180                                \\.RData$
#> 181                                  \\.RDS$
#> 182                               \\.sqlite$
#> 183                                   \\.h5$
#> 184                                  \\.md5$
#> 185                                \\.cksum$
#> 186                               \\.chksum$
#> 187                                  \\.gtf$
#> 188                                  \\.gff$
#> 189                                \\.fastq$
#> 190                                  \\.vcf$
#> 191                                \\.plink$
#> 192                                  \\.res$

pxFileTypes("foo")
#>   file type
#> 1  foo <NA>
pxFileTypes("foo.mzML")
#>       file type
#> 1 foo.mzML  raw
pxFileTypes("foo.raw")
#>      file   type
#> 1 foo.raw binary
pxFileTypes("foo.txt")
#>      file type
#> 1 foo.txt  doc
pxFileTypes("foo.R")
#>    file type
#> 1 foo.R code
pxFileTypes("foo.fasta")
#>        file type
#> 1 foo.fasta  fas

pxFileTypes(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
#>        file type
#> 1       foo <NA>
#> 2  foo.mzML  raw
#> 3     foo.R code
#> 4 foo.fasta  fas