Infer file type

The px_file_types() function inferres mass spectrometry and proteomics file types based on a currated table of file types and associated patterns. This table can be accessed with file_types(). See the examples below for the content of the tables.

The file types table is generated by scripts/make_file_types.R.

file_types()

px_file_types(fls, types = file_types())

Arguments

fls

character() of file names whose types need to be inferred based on their file extenstion.

types

data.frame of file types. Default is file_types().

Value

A data.frame with the filenames and their inferred types.

Author

Laurent Gatto with contributions from Dr. Samuel Wein to identify several file types.

Examples

file_types()
#>              type                                  ext
#> 1         archive                                  zip
#> 2         archive                               tar.gz
#> 3         archive                                  rar
#> 4         archive                                   7z
#> 5         archive                                  tgz
#> 6         archive                                  ZIP
#> 7         archive                                  tar
#> 8         archive                                  tgz
#> 9         archive                                  RAR
#> 10        archive                                  bz2
#> 11        archive                           webarchive
#> 12        archive                                   gz
#> 13        archive                                   xy
#> 14            doc                                  doc
#> 15            doc                                  pdf
#> 16            doc                                  PDF
#> 17            doc                                  ppt
#> 18            doc                                  odt
#> 19            doc                                 docx
#> 20            doc                                 pptx
#> 21            doc                                  rtf
#> 22            doc                                 html
#> 23            doc                                 html
#> 24            doc                                  htm
#> 25            doc                                shtml
#> 26            doc                               readme
#> 27            doc                                  txt
#> 28            doc                                   md
#> 29         binary                                  raw
#> 30         binary                                  Raw
#> 31         binary                                  RAW
#> 32         binary                                    d
#> 33         binary                                d.zip
#> 34         binary                              raw.zip
#> 35         binary                               raw.gz
#> 36         binary                                 wiff
#> 37         binary                                wiff2
#> 38         binary                            wiff.scan
#> 39         binary                         wiff.1.~idx2
#> 40         binary                             wiff.mtd
#> 41         binary                                  t2d
#> 42            raw                                 mzML
#> 43            raw                              mzML.gz
#> 44            raw                             mzML.zip
#> 45            raw                                mzXML
#> 46            raw                             mzXML.gz
#> 47            raw                            mzXML.zip
#> 48            raw                                mzxml
#> 49            raw                                TraML
#> 50            raw                                traML
#> 51            raw                                traml
#> 52            raw                               netCDF
#> 53            raw                               mzData
#> 54            raw                               mzdata
#> 55            raw                                  mz5
#> 56            raw                                imzML
#> 57            pkl                                  mgf
#> 58            pkl                                  MGF
#> 59            pkl                               mgf.gz
#> 60            pkl                               MGF.gz
#> 61            pkl                                  apl
#> 62            pkl                                  pkl
#> 63            pkl                               pkl.gz
#> 64            pkl                                  PKL
#> 65            fas                                  fas
#> 66            fas                                fasta
#> 67            fas                                   fa
#> 68            fas                                  faa
#> 69            fas                                FASTA
#> 70            fas                                fasts
#> 71            fas                             FALSE.gz
#> 72            fas                           FALSTA.zip
#> 73            fas                             fasta.gz
#> 74            fas                            fasta.zip
#> 75            fas                                fa.gz
#> 76            fas                               fa.zip
#> 77            fas                               faa.gz
#> 78            fas                              faa.zip
#> 79         reflib                                 blib
#> 80         reflib                                 elib
#> 81         reflib                                 dlib
#> 82         reflib                                  msp
#> 83             id                            mzIdentML
#> 84             id                                 mzID
#> 85             id                              mzID.gz
#> 86             id                                 mzid
#> 87             id                              mzid.gz
#> 88             id                             mzid.zip
#> 89             id                                  dat
#> 90             id                               dat.gz
#> 91             id                              dat.zip
#> 92             id                                idXML
#> 93             id                                  omx
#> 94             id                                  sqt
#> 95             id                                 pcml
#> 96             id                                  dta
#> 97             id                              dta.tgz
#> 98             id                          dta.tar.bz2
#> 99            tbl                                  csv
#> 100           tbl                                  tsv
#> 101           tbl                                  xls
#> 102           tbl                                 xlsx
#> 103           tbl                                 XLSX
#> 104           tbl                                 xlsb
#> 105           tbl                                  ssv
#> 106           tbl                               csv.gz
#> 107           tbl                               tsv.gz
#> 108           tbl                               psmtsv
#> 109           tbl                                delim
#> 110           tbl                              tabular
#> 111         mztab                                mztab
#> 112         mztab                             mztab.gz
#> 113         mztab                                mzTab
#> 114         mztab                             mzTab.gz
#> 115         mztab                              mzTabNA
#> 116         mztab                           -mztab.txt
#> 117         mztab                            mztab.txt
#> 118           fig                                  png
#> 119           fig                                  jpg
#> 120           fig                                 jpeg
#> 121           fig                                 tiff
#> 122           fig                                  TIF
#> 123           fig                                  tif
#> 124           fig                                  gif
#> 125           fig                                  PNG
#> 126           fig                                  JPG
#> 127           fig                                  svg
#> 128           xml                                  xml
#> 129           xml                               xml.gz
#> 130       prophet                               pepXML
#> 131       prophet                              protXML
#> 132        bruker                                  yep
#> 133        bruker                                  baf
#> 134      scaffold                                  sf3
#> 135      scaffold                                 sptm
#> 136      scaffold                                 sfdb
#> 137      scaffold                                 sdia
#> 138      scaffold                                metdb
#> 139            pd                             pdResult
#> 140            pd                                  msf
#> 141            pd                         pdResultView
#> 142            pd                              msfView
#> 143            pd                           pdAnalysis
#> 144            pd                       pdProcessingWF
#> 145            pd                        pdConsensusWF
#> 146            pd                              pdStudy
#> 147            pd                          pdStudy.bak
#> 148       sequest                                  ms1
#> 149       sequest                                  ms2
#> 150       sequest                                  srf
#> 151       sequest                                  out
#> 152       sequest                              out.tgz
#> 153       sequest                          out.tar.bz2
#> 154  proteinpilot                                group
#> 155    progenesis              ProgenesisQIPExperiment
#> 156    progenesis                 ProgenesisQIPArchive
#> 157    progenesis             ProgenesisLcmsExperiment
#> 158    progenesis ProgenesisQIPMultiFractionExperiment
#> 159       skyline                                  sky
#> 160       skyline                             sky.view
#> 161       skyline                                 skyd
#> 162       skyline                                 skyl
#> 163   spectronaut                                  sne
#> 164   spectronaut                                htrms
#> 165 peptideshaker                                 cpsx
#> 166        params                               params
#> 167        params                                 apar
#> 168        params                                 knwf
#> 169        params                                 json
#> 170          code                                    R
#> 171          code                                   py
#> 172          code                                    r
#> 173          code                                   pl
#> 174          code                                   js
#> 175          code                                  jar
#> 176          code                                  Rmd
#> 177          code                                   sh
#> 178          code                                ipynb
#> 179          data                                RData
#> 180          data                                  RDS
#> 181          data                               sqlite
#> 182          data                                   h5
#> 183           chk                                  md5
#> 184           chk                                cksum
#> 185           chk                               chksum
#> 186           gen                                  gtf
#> 187           gen                                  gff
#> 188           gen                                fastq
#> 189           gen                                  vcf
#> 190           gen                                plink
#> 191           gen                                  res
#>                                      pattern
#> 1                                    \\.zip$
#> 2                               \\.tar\\.gz$
#> 3                                    \\.rar$
#> 4                                     \\.7z$
#> 5                                    \\.tgz$
#> 6                                    \\.ZIP$
#> 7                                    \\.tar$
#> 8                                    \\.tgz$
#> 9                                    \\.RAR$
#> 10                                   \\.bz2$
#> 11                            \\.webarchive$
#> 12                                    \\.gz$
#> 13                                    \\.xy$
#> 14                                   \\.doc$
#> 15                                   \\.pdf$
#> 16                                   \\.PDF$
#> 17                                   \\.ppt$
#> 18                                   \\.odt$
#> 19                                  \\.docx$
#> 20                                  \\.pptx$
#> 21                                   \\.rtf$
#> 22                                  \\.html$
#> 23                                  \\.html$
#> 24                                   \\.htm$
#> 25                                 \\.shtml$
#> 26                                \\.readme$
#> 27                                   \\.txt$
#> 28                                    \\.md$
#> 29                                   \\.raw$
#> 30                                   \\.Raw$
#> 31                                   \\.RAW$
#> 32                                     \\.d$
#> 33                               \\.d\\.zip$
#> 34                             \\.raw\\.zip$
#> 35                              \\.raw\\.gz$
#> 36                                  \\.wiff$
#> 37                                 \\.wiff2$
#> 38                           \\.wiff\\.scan$
#> 39                      \\.wiff\\.1\\.~idx2$
#> 40                            \\.wiff\\.mtd$
#> 41                                   \\.t2d$
#> 42                                  \\.mzML$
#> 43                             \\.mzML\\.gz$
#> 44                            \\.mzML\\.zip$
#> 45                                 \\.mzXML$
#> 46                            \\.mzXML\\.gz$
#> 47                           \\.mzXML\\.zip$
#> 48                                 \\.mzxml$
#> 49                                 \\.TraML$
#> 50                                 \\.traML$
#> 51                                 \\.traml$
#> 52                                \\.netCDF$
#> 53                                \\.mzData$
#> 54                                \\.mzdata$
#> 55                                   \\.mz5$
#> 56                                 \\.imzML$
#> 57                                   \\.mgf$
#> 58                                   \\.MGF$
#> 59                              \\.mgf\\.gz$
#> 60                              \\.MGF\\.gz$
#> 61                                   \\.apl$
#> 62                                   \\.pkl$
#> 63                              \\.pkl\\.gz$
#> 64                                   \\.PKL$
#> 65                                   \\.fas$
#> 66                                 \\.fasta$
#> 67                                    \\.fa$
#> 68                                   \\.faa$
#> 69                                 \\.FASTA$
#> 70                                 \\.fasts$
#> 71                            \\.FALSE\\.gz$
#> 72                          \\.FALSTA\\.zip$
#> 73                            \\.fasta\\.gz$
#> 74                           \\.fasta\\.zip$
#> 75                               \\.fa\\.gz$
#> 76                              \\.fa\\.zip$
#> 77                              \\.faa\\.gz$
#> 78                             \\.faa\\.zip$
#> 79                                  \\.blib$
#> 80                                  \\.elib$
#> 81                                  \\.dlib$
#> 82                                   \\.msp$
#> 83                             \\.mzIdentML$
#> 84                                  \\.mzID$
#> 85                             \\.mzID\\.gz$
#> 86                                  \\.mzid$
#> 87                             \\.mzid\\.gz$
#> 88                            \\.mzid\\.zip$
#> 89                                   \\.dat$
#> 90                              \\.dat\\.gz$
#> 91                             \\.dat\\.zip$
#> 92                                 \\.idXML$
#> 93                                   \\.omx$
#> 94                                   \\.sqt$
#> 95                                  \\.pcml$
#> 96                                   \\.dta$
#> 97                             \\.dta\\.tgz$
#> 98                       \\.dta\\.tar\\.bz2$
#> 99                                   \\.csv$
#> 100                                  \\.tsv$
#> 101                                  \\.xls$
#> 102                                 \\.xlsx$
#> 103                                 \\.XLSX$
#> 104                                 \\.xlsb$
#> 105                                  \\.ssv$
#> 106                             \\.csv\\.gz$
#> 107                             \\.tsv\\.gz$
#> 108                               \\.psmtsv$
#> 109                                \\.delim$
#> 110                              \\.tabular$
#> 111                                \\.mztab$
#> 112                           \\.mztab\\.gz$
#> 113                                \\.mzTab$
#> 114                           \\.mzTab\\.gz$
#> 115                              \\.mzTabNA$
#> 116                            -mztab\\.txt$
#> 117                          \\.mztab\\.txt$
#> 118                                  \\.png$
#> 119                                  \\.jpg$
#> 120                                 \\.jpeg$
#> 121                                 \\.tiff$
#> 122                                  \\.TIF$
#> 123                                  \\.tif$
#> 124                                  \\.gif$
#> 125                                  \\.PNG$
#> 126                                  \\.JPG$
#> 127                                  \\.svg$
#> 128                                  \\.xml$
#> 129                             \\.xml\\.gz$
#> 130                               \\.pepXML$
#> 131                              \\.protXML$
#> 132                                  \\.yep$
#> 133                                  \\.baf$
#> 134                                  \\.sf3$
#> 135                                 \\.sptm$
#> 136                                 \\.sfdb$
#> 137                                 \\.sdia$
#> 138                                \\.metdb$
#> 139                             \\.pdResult$
#> 140                                  \\.msf$
#> 141                         \\.pdResultView$
#> 142                              \\.msfView$
#> 143                           \\.pdAnalysis$
#> 144                       \\.pdProcessingWF$
#> 145                        \\.pdConsensusWF$
#> 146                              \\.pdStudy$
#> 147                        \\.pdStudy\\.bak$
#> 148                                  \\.ms1$
#> 149                                  \\.ms2$
#> 150                                  \\.srf$
#> 151                                  \\.out$
#> 152                            \\.out\\.tgz$
#> 153                      \\.out\\.tar\\.bz2$
#> 154                                \\.group$
#> 155              \\.ProgenesisQIPExperiment$
#> 156                 \\.ProgenesisQIPArchive$
#> 157             \\.ProgenesisLcmsExperiment$
#> 158 \\.ProgenesisQIPMultiFractionExperiment$
#> 159                                  \\.sky$
#> 160                           \\.sky\\.view$
#> 161                                 \\.skyd$
#> 162                                 \\.skyl$
#> 163                                  \\.sne$
#> 164                                \\.htrms$
#> 165                                 \\.cpsx$
#> 166                               \\.params$
#> 167                                 \\.apar$
#> 168                                 \\.knwf$
#> 169                                 \\.json$
#> 170                                    \\.R$
#> 171                                   \\.py$
#> 172                                    \\.r$
#> 173                                   \\.pl$
#> 174                                   \\.js$
#> 175                                  \\.jar$
#> 176                                  \\.Rmd$
#> 177                                   \\.sh$
#> 178                                \\.ipynb$
#> 179                                \\.RData$
#> 180                                  \\.RDS$
#> 181                               \\.sqlite$
#> 182                                   \\.h5$
#> 183                                  \\.md5$
#> 184                                \\.cksum$
#> 185                               \\.chksum$
#> 186                                  \\.gtf$
#> 187                                  \\.gff$
#> 188                                \\.fastq$
#> 189                                  \\.vcf$
#> 190                                \\.plink$
#> 191                                  \\.res$

px_file_types("foo")
#>   file type
#> 1  foo <NA>
px_file_types("foo.mzML")
#>       file type
#> 1 foo.mzML  raw
px_file_types("foo.raw")
#>      file   type
#> 1 foo.raw binary
px_file_types("foo.txt")
#>      file type
#> 1 foo.txt  doc
px_file_types("foo.R")
#>    file type
#> 1 foo.R code
px_file_types("foo.fasta")
#>        file type
#> 1 foo.fasta  fas

px_file_types(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
#>        file type
#> 1       foo <NA>
#> 2  foo.mzML  raw
#> 3     foo.R code
#> 4 foo.fasta  fas