Class "PXDataset"

An S4 class to store and access information about ProteomeXchange (PX) data sets. Schema versions 1.0, 1.1 and 1.2 are supported and are documented on the PX code repository: https://code.google.com/p/proteomexchange/source/browse/schema/

Objects from the Class

Objects can be created with the constructor PXDataset.

Slots

id:

Object of class "character" storing the datasets unique identifier. Can be accessed with pxid().

formatVersion:

Object of class "character" storing the version of the ProteomeXchange schema.

Data:

Object of class "XMLNode" storing the ProteomeXchange description as XML node tree.

Details

From version 1.99.1 on, rpx uses the Bioconductor BiocFileCache package to automatically cache all downloaded ProteomeXchange files. When a file is downloaded for the first time, it is added to the cache. When already available, the file path to the cached file is returned. The rpx chache is returned by rpx:::.get_cache(). The user is asked to confirm its creation when data is first downloaded.

For more details on how to manage the cache (for example if some files need to be deleted), please refer to the BiocFileCache vignette.

Methods

pxfiles

signature(object = "PXDataset"): return a character of all available files.

pxget

signature(object = "PXDataset", list, ...): downloads the files from the ProteomeXchange repository. If list is missing, the file to be downloaded can be selected from a menu. If list = "all", all files are downloaded. The file names, as returned by pxfiles can also be used. Alternatively, a logical or numeric indices can be used. If not already cached, the files are downloaded and added to the package cache. The function then returns the names of the files in the cache directory. Additional argument cache can be passed to the internal function rpx:::pxget1() that downloads files to and retrieves them from cache. The default cache is the packages' default rappdirs::user_cache_dir(appname = "rpx").

pxid

signature(object = "PXDataset"): returns the unique ProteomeXchange identifier.

pxref

signature(object = "PXDataset"): returns the reference associcated with the object.

pxtax

signature(object = "PXDataset"): returns the scientic taxonomic name of object.

pxurl

signature(object = "PXDataset": returns the base url on the ProteomeXchange server where pxfiles(object) reside.

show

signature(object = "PXDataset"): textual representation of object.

References

Vizcaino J.A. et al. 'ProteomeXchange: globally co-ordinated proteomics data submission and dissemination', Nature Biotechnology 2014, 32, 223 -- 226, doi:10.1038/nbt.2839.

Source repository for the ProteomeXchange project: https://code.google.com/p/proteomexchange/

Author

Laurent Gatto

Examples

px <- PXDataset("PXD000001")
px
#> Object of class "PXDataset"
#>  Id: PXD000001 with 11 files
#>  [1] 'F063721.dat' ... [11] 'erwinia_carotovora.fasta'
#>  Use 'pxfiles(.)' to see all files.
pxtax(px)
#> [1] "Erwinia carotovora"
pxurl(px)
#> [1] "ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2012/03/PXD000001"
pxref(px)
#> [1] "Gatto L, Christoforou A. Using R and Bioconductor for proteomics data analysis. Biochim Biophys Acta. 2013 May 18. doi:pii: S1570-9639(13)00186-6. 10.1016/j.bbapap.2013.04.032"
pxfiles(px)
#>  [1] "F063721.dat"                                                         
#>  [2] "F063721.dat-mztab.txt"                                               
#>  [3] "PRIDE_Exp_Complete_Ac_22134.xml.gz"                                  
#>  [4] "PRIDE_Exp_mzData_Ac_22134.xml.gz"                                    
#>  [5] "PXD000001_mztab.txt"                                                 
#>  [6] "README.txt"                                                          
#>  [7] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML" 
#>  [8] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzXML"
#>  [9] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML"         
#> [10] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.raw"           
#> [11] "erwinia_carotovora.fasta"                                            

fas <- pxget(px, "erwinia_carotovora.fasta")
#> Loading erwinia_carotovora.fasta from cache.
fas
#> [1] "/github/home/.cache/R/rpx/e051f678efb_erwinia_carotovora.fasta"
library("Biostrings")
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: XVector
#> Loading required package: GenomeInfoDb
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
readAAStringSet(fas)
#> AAStringSet object of length 4499:
#>        width seq                                            names               
#>    [1]   147 MADITLISGSTLGSAEYVAEHL...QHQIPEDPAEEWLGSWVNLLK ECA0001 putative ...
#>    [2]   153 VAEIYQIDNLDRGILSALMENA...EIQSTETLISLQNPIMRTIAP ECA0002 AsnC-fami...
#>    [3]   330 MKKQYIEKQQQISFVKSFFSSQ...IGQVQCGVWPQPLRESVSGLL ECA0003 putative ...
#>    [4]   492 MITLESLEMLLSIDENELLDDL...WRFDTGLKSRLMRRWQHGKAY ECA0004 conserved...
#>    [5]   499 MRQTAALAERISRLSHALEHGL...AKIEASLQQVAEQIQQSEQQD ECA0005 conserved...
#>    ...   ... ...
#> [4495]   634 MSDKIIHLTDDSFDTDVLKADG...RRKVDPLRVFASDMARRLELL trx-rv3790 trx-rv...
#> [4496]    93 MTKMNNKARRTARELKHLGASI...RELRDEFPMGYLGDYKDDDDK TimBlower TimBlower
#> [4497]   309 MFSNLSKRWAQRTLSKSFYSTA...KFKWAGIKTRKFVFNPPKPRK sp|P07143|CY1_YEA...
#> [4498]   231 FPTDDDDKIVGGYTCAANSIPY...PGVYTKVCNYVNWIQQTIAAN sp|P00761|TRYP_PI...
#> [4499]   269 GVSGSCNIDVVCPEGNGHRDVI...DAAGTGAQFIDGLDSTGTPPV sp|Q7M135|LYSC_LY...