Class

An S4 class to store and access information about ProteomeXchange (PX) data sets. Schema versions 1.0, 1.1 and 1.2 are supported and are documented on the PX code repository: https://code.google.com/p/proteomexchange/source/browse/schema/

Objects from the Class

Objects can be created with the constructor PXDataset.

Slots

id:

Object of class "character" storing the datasets unique identifier. Can be accessed with pxid().

formatVersion:

Object of class "character" storing the version of the ProteomeXchange schema.

Data:

Object of class "XMLNode" storing the ProteomeXchange description as XML node tree.

Methods

pxfiles

signature(object = "PXDataset"): return a character of all available files.

pxget

signature(object = "PXDataset", list, force = FALSE, ...): downloads the files from the ProteomeXchange repository. If list is missing, the file to be downloaded can be selected from a menu. If list = "all", all files are downloaded. The file names, as returned by pxfiles can also be used. Alternatively, a logical or numeric indices can be used. All files will be downloaded in the working directory. Unless force is set to TRUE, files are not downloaded if already present in the working directory. Additional parameters can be passed to download.file via .... Invisibly returns the names of the downloaded files.

pxid

signature(object = "PXDataset"): returns the unique ProteomeXchange identifier.

pxref

signature(object = "PXDataset"): returns the reference associcated with the object.

pxtax

signature(object = "PXDataset"): returns the scientic taxonomic name of object.

pxurl

signature(object = "PXDataset"): returns the base url on the ProteomeXchange server where pxfiles(object) reside.

show

signature(object = "PXDataset"): textual representation of object.

References

Vizcaino J.A. et al. 'ProteomeXchange: globally co-ordinated proteomics data submission and dissemination', Nature Biotechnology 2014, 32, 223 -- 226, doi:10.1038/nbt.2839.

Source repository for the ProteomeXchange project: https://code.google.com/p/proteomexchange/

Examples

px <- PXDataset("PXD000001")
px

#> Object of class "PXDataset"
#>  Id: PXD000001 with 10 files
#>  [1] 'F063721.dat' ... [10] 'erwinia_carotovora.fasta'
#>  Use 'pxfiles(.)' to see all files.

pxtax(px)

#> [1] "Erwinia carotovora"

pxurl(px)

#> [1] "ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2012/03/PXD000001"

pxref(px)

#> [1] "Gatto L, Christoforou A. Using R and Bioconductor for proteomics data analysis. Biochim Biophys Acta. 2014 Jan;1844(1 Pt A):42-51. Review"

pxfiles(px)

#>  [1] "F063721.dat"                                                         
#>  [2] "F063721.dat-mztab.txt"                                               
#>  [3] "PRIDE_Exp_Complete_Ac_22134.xml.gz"                                  
#>  [4] "PRIDE_Exp_mzData_Ac_22134.xml.gz"                                    
#>  [5] "PXD000001_mztab.txt"                                                 
#>  [6] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML" 
#>  [7] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzXML"
#>  [8] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML"         
#>  [9] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.raw"           
#> [10] "erwinia_carotovora.fasta"                                            

fnm <- pxget(px, "PXD000001_mztab.txt")

#> Downloading 1 file

library("MSnbase")

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, cbind, colMeans, colnames,
#>     colSums, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, lengths, Map, mapply, match,
#>     mget, order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rowMeans, rownames, rowSums, sapply, setdiff, sort,
#>     table, tapply, union, unique, unsplit, which, which.max, which.min

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: mzR

#> Loading required package: Rcpp

#> Loading required package: BiocParallel

#> Loading required package: ProtGenerics

#> 
#> This is MSnbase version 2.1.5 
#>   Read '?MSnbase' and references therein for information
#>   about the package and how to get started.

#> 
#> Attaching package: ‘MSnbase’

#> The following object is masked from ‘package:stats’:
#> 
#>     smooth

#> The following object is masked from ‘package:base’:
#> 
#>     trimws

readMzTabData(fnm, "PEP")

#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 2351 features, 0 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData: none
#> featureData
#>   featureNames: FGNDAR DGVSVAR ... LPALLK (2351 total)
#>   fvarLabels: sequence accession ... peptide_abundance_std_error_sub.6.
#>     (33 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#>  MSnbase version: 2.1.5 

unlink("PXD000001_mztab.txt")