New PXDataset (v2) to find and download proteomics data

The rpx package provides the infrastructure to access, store and retrieve information for ProteomeXchange (PX) data sets. This can be achieved with PXDataset2 objects can be created with the PXDataset2() constructor that takes the unique ProteomeXchange project identifier as input.

The new PXDataset2 class superseeds the previous and now deprecated PXDataset version.

PXDataset2(id, cache = rpxCache())

PXDataset(id, cache = rpxCache())

# S4 method for PXDataset2
pxid(object)

# S4 method for PXDataset2
pxurl(object)

# S4 method for PXDataset2
pxtax(object)

# S4 method for PXDataset2
pxref(object)

pxtitle(object)

# S4 method for PXDataset2
pxfiles(object, n = 10)

# S4 method for PXDataset2
pxCacheInfo(object)

# S4 method for PXDataset2
pxget(object, list, cache = rpxCache())

Arguments

id
character(1) containing a valid ProteomeXchange identifier.

cache
Object of class BiocFileCache. Default is to use the central rpx cache returned by rpxCache(), but users can use their own cache. See rpxCache() for details.

object
An instance of class PXDataset2.

n
integer(1) indicating the number of files to be printed.

list
character(), numeric() or logical() defining the project files to be downloaded. This list of files can retrieved with pxfiles().

Value

The PXDataset2() returns a cached PXDataset2 object. It thus also modifies the cache used to projet caching, as defined by the cache argument.

Details

The rpx packages uses caching to store ProteomeXchange projects and project files. When creating an object with PXDataset2(), the cache is first queried for the projects identifier. If a unique hit is found, the project is retrieved and returned. If no matching project identifier is found, then the remote resource is accessed to first create the new PXDataset2() project, then cache it before returning it to the user. The same mechanism is applied when project files are requested.

Caching is supported by BiocFileCache package. The PXDataset2() constructor and the px_get() function can be passed a instance of class BiocFileCache that defines the cache. The default is to use the package-wide cache defined in rpxCache(). For more details on how to manage the cache (for example if some files need to be deleted), please refer to the BiocFileCache package vignette and documentation. See also rpxCache() for additional details.

Slots

px_id

character(1) containing the dataset's unique ProteomeXchange identifier, as used to create the object.

px_rid

character(1) storing the cached resource name in the BiocFileCache instance stored in cachepath.

px_title

character(1) with the project's title.

px_url

`character(1) with the project's URL.

px_doi

character(1) with the project's DOI.

px_ref

character containing the project's reference(s).

px_ref_doi

character containing the project's reference DOIs.

px_pubmed

character containing the project's reference PubMed identifier.

px_files

data.frame containing information about the project files, including file names, URIs and types. The files are retrieved from the project's README.txt file.

px_tax

charcter (typically of length 1) containing the taxonomy of the sample.

px_metadata

list containing the project's metadata, as downloaded from the ProteomeXchange site. All slots but px_files are populated from this one.

cachepath

character(1) storing the path to the cache the project object is stored in.

Accessors

pxfiles(object, n = 10) invisibly returns all the project file names. The function prints the first n files specifying whether they are local of remote (based on the cache the object is stored in). The printing can be ignored by wrapping the call in suppressMessages().

pxget(object, list, cache): list is a vector defining the files to be downloaded. If list = "all", all files are downloaded. The file names, as returned by pxfiles() can also be used. Alternatively, a logical or numeric index can be used. If missing, the file to be downloaded can be selected from a menu.

The argument cache can be passed to define the path to the cache. The default cache is the packages' default as returned by rpxCache().

pxtax(object): returns the taxonomic name of object.

pxurl(object): returns the base url on the ProteomeXchange server where the project files reside.

pxCacheInfo(object, cache): prints and invisibly returns object's caching information from cache(default isrpxCache()`). The return value is a named vector of length two containing the resourne identifier and the cache location.

`pxtitle(object): returns the project's title.

pxref(object): returns the project's bibliographic reference(s).

References

Vizcaino J.A. et al. 'ProteomeXchange: globally co-ordinated proteomics data submission and dissemination', Nature Biotechnology 2014, 32, 223 -- 226, doi:10.1038/nbt.2839.

Source repository for the ProteomeXchange project: https://code.google.com/p/proteomexchange/

Author

Laurent Gatto

Examples

px <- PXDataset2("PXD000001")
#> Loading PXD000001 from cache.
px
#> Project PXD000001 with 10 files
#>  
#> Resource ID BFC1 in cache in /github/home/.cache/R/rpx.
#>  [1] 'erwinia_carotovora.fasta' ... [10] 'PRIDE_Exp_Complete_Ac_22134.pride.mgf.gz'
#>  Use 'pxfiles(.)' to see all files.
pxtax(px)
#> [1] "Erwinia carotovora"
pxurl(px)
#> [1] "ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2012/03/PXD000001"
pxref(px)
#> [1] "Gatto L, Christoforou A; Using R and Bioconductor for proteomics data analysis., Biochim Biophys Acta, 2013 May 18, doi:10.1016/j.bbapap.2013.04.032 PMID:23692960"
pxfiles(px)
#> Project PXD000001 files (10):
#>  [local]  erwinia_carotovora.fasta
#>  [remote] F063721.dat
#>  [local]  F063721.dat-mztab.txt
#>  [remote] PRIDE_Exp_Complete_Ac_22134.xml.gz
#>  [remote] TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.raw
#>  [remote] TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML
#>  [remote] TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzXML
#>  [remote] TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML
#>  [remote] PRIDE_Exp_Complete_Ac_22134.pride.mztab.gz
#>  [remote] PRIDE_Exp_Complete_Ac_22134.pride.mgf.gz
pxCacheInfo(px)
#> Resource ID BFC1 in cache in /github/home/.cache/R/rpx.

fas <- pxget(px, "erwinia_carotovora.fasta")
#> Loading erwinia_carotovora.fasta from cache.
fas
#> [1] "/github/home/.cache/R/rpx/5e7462bb17c_erwinia_carotovora.fasta"
library("Biostrings")
readAAStringSet(fas)
#> AAStringSet object of length 4499:
#>        width seq                                            names               
#>    [1]   147 MADITLISGSTLGSAEYVAEHL...QHQIPEDPAEEWLGSWVNLLK ECA0001 putative ...
#>    [2]   153 VAEIYQIDNLDRGILSALMENA...EIQSTETLISLQNPIMRTIAP ECA0002 AsnC-fami...
#>    [3]   330 MKKQYIEKQQQISFVKSFFSSQ...IGQVQCGVWPQPLRESVSGLL ECA0003 putative ...
#>    [4]   492 MITLESLEMLLSIDENELLDDL...WRFDTGLKSRLMRRWQHGKAY ECA0004 conserved...
#>    [5]   499 MRQTAALAERISRLSHALEHGL...AKIEASLQQVAEQIQQSEQQD ECA0005 conserved...
#>    ...   ... ...
#> [4495]   634 MSDKIIHLTDDSFDTDVLKADG...RRKVDPLRVFASDMARRLELL trx-rv3790 trx-rv...
#> [4496]    93 MTKMNNKARRTARELKHLGASI...RELRDEFPMGYLGDYKDDDDK TimBlower TimBlower
#> [4497]   309 MFSNLSKRWAQRTLSKSFYSTA...KFKWAGIKTRKFVFNPPKPRK sp|P07143|CY1_YEA...
#> [4498]   231 FPTDDDDKIVGGYTCAANSIPY...PGVYTKVCNYVNWIQQTIAAN sp|P00761|TRYP_PI...
#> [4499]   269 GVSGSCNIDVVCPEGNGHRDVI...DAAGTGAQFIDGLDSTGTPPV sp|Q7M135|LYSC_LY...