Package cache

Function to access and manage the cache. rxpCache() returns the central rpx cache. pxCachedrojects() prints the names of the cached projects and invisibly returns the cache table.

rpxCache()

pxCachedProjects(cache = rpxCache(), rpxprefix = "^\\.rpx(2?)")

Arguments

cache
Object of class BiocFileCache.

rpxprefix
character(1) defining the resourne name prefix in cache. Default is "^\\.rpx(2?)" to match objects of class PXDataset and PXDataset2.

Value

The rpxCache() function returns an instance of class BiocFileCache. pxCachedProjects() invisbly returns a tibble of cached ProteomeXchange projects.

Details

The cache is an object of class BiocFileCache, and created with BiocFileCache::BiocFileCache(). It can be either the package-wide cache as defined by rpxCache() or an instaned provided by the user.

When projects are cached, they are given a resource name (rname) composed of the .rpx prefix followed by the ProteomeXchange identifier. For example, project PXD000001 is named .rpxPXD000001 to avoid any conflicts with other resources that user-created resources.

Author

Laurent Gatto

Examples

## Default rpx cache
rpxCache()
#> class: BiocFileCache
#> bfccache: /github/home/.cache/R/rpx
#> bfccount: 4
#> For more information see: bfcinfo() or bfcquery()

## Set up your own cache by providing a file or a directory to
## BiocFileCache::BiocFileCache()
my_cache <- BiocFileCache::BiocFileCache(tempfile())
my_cache
#> class: BiocFileCache
#> bfccache: /tmp/RtmpYZQlyV/file9561012ff2e
#> bfccount: 0
#> For more information see: bfcinfo() or bfcquery()
if (FALSE) {
px <- PXDataset2("PXD000001", cache = my_cache)
pxget(px, "README.txt", cache = my_cache)
}

## List of cached projects
pxCachedProjects()
#> Cached projects (2): PXD000001, PXD022816 

## To delete project a project from the default cache, first find
## its resource id (rid) in the cache
(cache_tbl <- pxCachedProjects())
#> Cached projects (2): PXD000001, PXD022816 
#> # A tibble: 2 × 10
#>   rid   rname  create_time access_time rpath  rtype fpath last_modified_t… etag 
#>   <chr> <chr>  <chr>       <chr>       <chr>  <chr> <chr>            <dbl> <chr>
#> 1 BFC1  .rpx2… 2021-08-16… 2021-08-16… /gith… rela… 5e77…               NA NA   
#> 2 BFC4  .rpx2… 2021-08-16… 2021-08-16… /gith… rela… 7ab1…               NA NA   
#> # … with 1 more variable: expires <dbl>
(rid <- cache_tbl[cache_tbl$rname == ".rpxPXD000001", "rid"][[1]])
#> character(0)

## Alternatively, extact the information from the project
px <- PXDataset2("PXD000001")
#> Loading PXD000001 from cache.
px1_cache_info <- pxCacheInfo(px)
#> Resource ID BFC1 in cache in /github/home/.cache/R/rpx.
(rid <- px1_cache_info["rid"])
#>    rid 
#> "BFC1" 

## Then remove it with BiocFileCache:: bfcremove()
if (FALSE) {
BiocFileCache:::bfcremove(rpxCache(), rid)
}