Infer file type

The pxFileTypes() function inferres mass spectrometry and proteomics file types based on a currated table of file types and associated patterns. This table can be accessed with fileTypes(). See the examples below for the content and format of the table.

The types of the files in a PXDataset object can be accessed with the pxfiles(as.vector = FALSE) function. See examples in the pxfiles() manual page.

updatePxFileTypes() updates the file types of a PXDataset instance using pxFileTypes(). This function also udpates the cached object unless cache is set to NULL. This function is useful to harmonise file types when the data in fileTypes() is updated.

The file types table is generated by scripts/make_fileTypes.R.

fileTypes()

pxFileTypes(fls, types = fileTypes())

updatePxFileTypes(object, cache = rpxCache())

Arguments

fls

character() of file names whose types need to be inferred based on their file extenstion.

types

data.frame of file types. Default is fileTypes().

object

Object of class PXDataset.

cache

Object of class BiocFileCache.

Value

A data.frame with the filenames and their inferred types.

Author

Laurent Gatto with contributions via mastodon from Dr. Samuel Wein and Michael MacCoss to identify several file types.

Examples

fileTypes()
#>              type                                  ext
#> 1         archive                                  zip
#> 2         archive                               tar.gz
#> 3         archive                                  rar
#> 4         archive                                  RAR
#> 5         archive                                   7z
#> 6         archive                                  tgz
#> 7         archive                                  ZIP
#> 8         archive                                  tar
#> 9         archive                                  tgz
#> 10        archive                                  RAR
#> 11        archive                                  bz2
#> 12        archive                           webarchive
#> 13        archive                                   gz
#> 14        archive                                   xy
#> 15            doc                                  doc
#> 16            doc                                  pdf
#> 17            doc                                  PDF
#> 18            doc                                  ppt
#> 19            doc                                  odt
#> 20            doc                                 docx
#> 21            doc                                 pptx
#> 22            doc                                  rtf
#> 23            doc                                 html
#> 24            doc                                 html
#> 25            doc                                  htm
#> 26            doc                                shtml
#> 27            doc                               readme
#> 28            doc                                  txt
#> 29            doc                                   md
#> 30            doc                                  css
#> 31         rawbin                                  raw
#> 32         rawbin                                  Raw
#> 33         rawbin                                  RAW
#> 34         rawbin                                    d
#> 35         rawbin                                d.zip
#> 36         rawbin                              raw.zip
#> 37         rawbin                               raw.gz
#> 38         rawbin                                 wiff
#> 39         rawbin                                wiff2
#> 40         rawbin                            wiff.scan
#> 41         rawbin                         wiff.1.~idx2
#> 42         rawbin                             wiff.mtd
#> 43         rawbin                                  t2d
#> 44            raw                                 mzML
#> 45            raw                              mzML.gz
#> 46            raw                             mzML.zip
#> 47            raw                                mzXML
#> 48            raw                             mzXML.gz
#> 49            raw                            mzXML.zip
#> 50            raw                                mzxml
#> 51            raw                                MZXML
#> 52            raw                                TraML
#> 53            raw                                traML
#> 54            raw                                traml
#> 55            raw                               netCDF
#> 56            raw                                  CDF
#> 57            raw                               mzData
#> 58            raw                               mzdata
#> 59            raw                                  mz5
#> 60            raw                                imzML
#> 61            pkl                                  mgf
#> 62            pkl                                  MGF
#> 63            pkl                               mgf.gz
#> 64            pkl                               MGF.gz
#> 65            pkl                                  apl
#> 66            pkl                                  pkl
#> 67            pkl                               pkl.gz
#> 68            pkl                                  PKL
#> 69            fas                                  fas
#> 70            fas                                fasta
#> 71            fas                                   fa
#> 72            fas                                  faa
#> 73            fas                                FASTA
#> 74            fas                                fasts
#> 75            fas                             FALSE.gz
#> 76            fas                           FALSTA.zip
#> 77            fas                             fasta.gz
#> 78            fas                            fasta.zip
#> 79            fas                                fa.gz
#> 80            fas                               fa.zip
#> 81            fas                               faa.gz
#> 82            fas                              faa.zip
#> 83         reflib                                 blib
#> 84         reflib                                 elib
#> 85         reflib                                 dlib
#> 86         reflib                                  msp
#> 87             id                            mzIdentML
#> 88             id                            mzidentml
#> 89             id                            mzidentML
#> 90             id                                 mzID
#> 91             id                              mzID.gz
#> 92             id                                 mzid
#> 93             id                              mzid.gz
#> 94             id                             mzid.zip
#> 95             id                                  dat
#> 96             id                               dat.gz
#> 97             id                              dat.zip
#> 98             id                                idXML
#> 99             id                                  omx
#> 100            id                                IdXML
#> 101            id                                idxml
#> 102            id                              pepnovo
#> 103            id                                 pcml
#> 104            id                                  dta
#> 105            id                              dta.tgz
#> 106            id                          dta.tar.bz2
#> 107           tbl                                  csv
#> 108           tbl                                  tsv
#> 109           tbl                                  xls
#> 110           tbl                                 xlsx
#> 111           tbl                                 XLSX
#> 112           tbl                                 xlsb
#> 113           tbl                                  ssv
#> 114           tbl                               csv.gz
#> 115           tbl                               tsv.gz
#> 116           tbl                               psmtsv
#> 117           tbl                                delim
#> 118           tbl                              tabular
#> 119         mztab                                mztab
#> 120         mztab                             mztab.gz
#> 121         mztab                                mzTab
#> 122         mztab                             mzTab.gz
#> 123         mztab                              mzTabNA
#> 124         mztab                           -mztab.txt
#> 125         mztab                           _mztab.txt
#> 126         mztab                            mztab.txt
#> 127           fig                                  png
#> 128           fig                                  jpg
#> 129           fig                                 jpeg
#> 130           fig                                 tiff
#> 131           fig                                  TIF
#> 132           fig                                  tif
#> 133           fig                                  gif
#> 134           fig                                  PNG
#> 135           fig                                  JPG
#> 136           fig                                  svg
#> 137           xml                                  xml
#> 138           xml                               xml.gz
#> 139       prophet                               pepXML
#> 140       prophet                              protXML
#> 141       prophet                               pepxml
#> 142       prophet                              protxml
#> 143        bruker                                  yep
#> 144        bruker                                  baf
#> 145      scaffold                                  sf3
#> 146      scaffold                                 sptm
#> 147      scaffold                                 sfdb
#> 148      scaffold                                 sdia
#> 149      scaffold                                metdb
#> 150            pd                             pdResult
#> 151            pd                                  msf
#> 152            pd                         pdResultView
#> 153            pd                              msfView
#> 154            pd                           pdAnalysis
#> 155            pd                       pdProcessingWF
#> 156            pd                        pdConsensusWF
#> 157            pd                              pdStudy
#> 158            pd                          pdStudy.bak
#> 159       sequest                                  ms1
#> 160       sequest                                  ms2
#> 161       sequest                                  srf
#> 162       sequest                                  sqt
#> 163       sequest                                  out
#> 164       sequest                              out.tgz
#> 165       sequest                          out.tar.bz2
#> 166  proteinpilot                                group
#> 167    progenesis              ProgenesisQIPExperiment
#> 168    progenesis                 ProgenesisQIPArchive
#> 169    progenesis             ProgenesisLcmsExperiment
#> 170    progenesis ProgenesisQIPMultiFractionExperiment
#> 171       skyline                                  sky
#> 172       skyline                             sky.view
#> 173       skyline                                 view
#> 174       skyline                                 skyd
#> 175       skyline                                 skyl
#> 176   spectronaut                                  sne
#> 177   spectronaut                                htrms
#> 178 peptideshaker                                 cpsx
#> 179        params                               PARAMS
#> 180        params                               params
#> 181        params                                param
#> 182        params                                  par
#> 183        params                               config
#> 184        params                                 apar
#> 185        params                                 knwf
#> 186        params                                 json
#> 187        params                                 toml
#> 188        params                                 yaml
#> 189        params                                  ini
#> 190        params                                  mtd
#> 191        params                                index
#> 192        params                               method
#> 193        params                               Method
#> 194        params                             FAmethod
#> 195        params                           properties
#> 196          code                                    R
#> 197          code                                   py
#> 198          code                                    r
#> 199          code                                   pl
#> 200          code                                   js
#> 201          code                                  jar
#> 202          code                                  Rmd
#> 203          code                                   sh
#> 204          code                                ipynb
#> 205           exe                                  exe
#> 206           exe                                  bin
#> 207           exe                                  dll
#> 208          data                                RData
#> 209          data                                  RDS
#> 210          data                               sqlite
#> 211          data                                   h5
#> 212           chk                                  md5
#> 213           chk                                cksum
#> 214           chk                               chksum
#> 215           tmp                                  bak
#> 216           tmp                             download
#> 217           tmp                           crdownload
#> 218           tmp                           sgdownload
#> 219           tmp                                 temp
#> 220           tmp                                  tmp
#> 221           gen                                  gtf
#> 222           gen                                  gff
#> 223           gen                                fastq
#> 224           gen                                  vcf
#> 225           gen                                plink
#> 226           gen                                  res
#>                                      pattern
#> 1                                    \\.zip$
#> 2                               \\.tar\\.gz$
#> 3                                    \\.rar$
#> 4                                    \\.RAR$
#> 5                                     \\.7z$
#> 6                                    \\.tgz$
#> 7                                    \\.ZIP$
#> 8                                    \\.tar$
#> 9                                    \\.tgz$
#> 10                                   \\.RAR$
#> 11                                   \\.bz2$
#> 12                            \\.webarchive$
#> 13                                    \\.gz$
#> 14                                    \\.xy$
#> 15                                   \\.doc$
#> 16                                   \\.pdf$
#> 17                                   \\.PDF$
#> 18                                   \\.ppt$
#> 19                                   \\.odt$
#> 20                                  \\.docx$
#> 21                                  \\.pptx$
#> 22                                   \\.rtf$
#> 23                                  \\.html$
#> 24                                  \\.html$
#> 25                                   \\.htm$
#> 26                                 \\.shtml$
#> 27                                \\.readme$
#> 28                                   \\.txt$
#> 29                                    \\.md$
#> 30                                   \\.css$
#> 31                                   \\.raw$
#> 32                                   \\.Raw$
#> 33                                   \\.RAW$
#> 34                                     \\.d$
#> 35                               \\.d\\.zip$
#> 36                             \\.raw\\.zip$
#> 37                              \\.raw\\.gz$
#> 38                                  \\.wiff$
#> 39                                 \\.wiff2$
#> 40                           \\.wiff\\.scan$
#> 41                      \\.wiff\\.1\\.~idx2$
#> 42                            \\.wiff\\.mtd$
#> 43                                   \\.t2d$
#> 44                                  \\.mzML$
#> 45                             \\.mzML\\.gz$
#> 46                            \\.mzML\\.zip$
#> 47                                 \\.mzXML$
#> 48                            \\.mzXML\\.gz$
#> 49                           \\.mzXML\\.zip$
#> 50                                 \\.mzxml$
#> 51                                 \\.MZXML$
#> 52                                 \\.TraML$
#> 53                                 \\.traML$
#> 54                                 \\.traml$
#> 55                                \\.netCDF$
#> 56                                   \\.CDF$
#> 57                                \\.mzData$
#> 58                                \\.mzdata$
#> 59                                   \\.mz5$
#> 60                                 \\.imzML$
#> 61                                   \\.mgf$
#> 62                                   \\.MGF$
#> 63                              \\.mgf\\.gz$
#> 64                              \\.MGF\\.gz$
#> 65                                   \\.apl$
#> 66                                   \\.pkl$
#> 67                              \\.pkl\\.gz$
#> 68                                   \\.PKL$
#> 69                                   \\.fas$
#> 70                                 \\.fasta$
#> 71                                    \\.fa$
#> 72                                   \\.faa$
#> 73                                 \\.FASTA$
#> 74                                 \\.fasts$
#> 75                            \\.FALSE\\.gz$
#> 76                          \\.FALSTA\\.zip$
#> 77                            \\.fasta\\.gz$
#> 78                           \\.fasta\\.zip$
#> 79                               \\.fa\\.gz$
#> 80                              \\.fa\\.zip$
#> 81                              \\.faa\\.gz$
#> 82                             \\.faa\\.zip$
#> 83                                  \\.blib$
#> 84                                  \\.elib$
#> 85                                  \\.dlib$
#> 86                                   \\.msp$
#> 87                             \\.mzIdentML$
#> 88                             \\.mzidentml$
#> 89                             \\.mzidentML$
#> 90                                  \\.mzID$
#> 91                             \\.mzID\\.gz$
#> 92                                  \\.mzid$
#> 93                             \\.mzid\\.gz$
#> 94                            \\.mzid\\.zip$
#> 95                                   \\.dat$
#> 96                              \\.dat\\.gz$
#> 97                             \\.dat\\.zip$
#> 98                                 \\.idXML$
#> 99                                   \\.omx$
#> 100                                \\.IdXML$
#> 101                                \\.idxml$
#> 102                              \\.pepnovo$
#> 103                                 \\.pcml$
#> 104                                  \\.dta$
#> 105                            \\.dta\\.tgz$
#> 106                      \\.dta\\.tar\\.bz2$
#> 107                                  \\.csv$
#> 108                                  \\.tsv$
#> 109                                  \\.xls$
#> 110                                 \\.xlsx$
#> 111                                 \\.XLSX$
#> 112                                 \\.xlsb$
#> 113                                  \\.ssv$
#> 114                             \\.csv\\.gz$
#> 115                             \\.tsv\\.gz$
#> 116                               \\.psmtsv$
#> 117                                \\.delim$
#> 118                              \\.tabular$
#> 119                                \\.mztab$
#> 120                           \\.mztab\\.gz$
#> 121                                \\.mzTab$
#> 122                           \\.mzTab\\.gz$
#> 123                              \\.mzTabNA$
#> 124                            -mztab\\.txt$
#> 125                            _mztab\\.txt$
#> 126                          \\.mztab\\.txt$
#> 127                                  \\.png$
#> 128                                  \\.jpg$
#> 129                                 \\.jpeg$
#> 130                                 \\.tiff$
#> 131                                  \\.TIF$
#> 132                                  \\.tif$
#> 133                                  \\.gif$
#> 134                                  \\.PNG$
#> 135                                  \\.JPG$
#> 136                                  \\.svg$
#> 137                                  \\.xml$
#> 138                             \\.xml\\.gz$
#> 139                               \\.pepXML$
#> 140                              \\.protXML$
#> 141                               \\.pepxml$
#> 142                              \\.protxml$
#> 143                                  \\.yep$
#> 144                                  \\.baf$
#> 145                                  \\.sf3$
#> 146                                 \\.sptm$
#> 147                                 \\.sfdb$
#> 148                                 \\.sdia$
#> 149                                \\.metdb$
#> 150                             \\.pdResult$
#> 151                                  \\.msf$
#> 152                         \\.pdResultView$
#> 153                              \\.msfView$
#> 154                           \\.pdAnalysis$
#> 155                       \\.pdProcessingWF$
#> 156                        \\.pdConsensusWF$
#> 157                              \\.pdStudy$
#> 158                        \\.pdStudy\\.bak$
#> 159                                  \\.ms1$
#> 160                                  \\.ms2$
#> 161                                  \\.srf$
#> 162                                  \\.sqt$
#> 163                                  \\.out$
#> 164                            \\.out\\.tgz$
#> 165                      \\.out\\.tar\\.bz2$
#> 166                                \\.group$
#> 167              \\.ProgenesisQIPExperiment$
#> 168                 \\.ProgenesisQIPArchive$
#> 169             \\.ProgenesisLcmsExperiment$
#> 170 \\.ProgenesisQIPMultiFractionExperiment$
#> 171                                  \\.sky$
#> 172                           \\.sky\\.view$
#> 173                                 \\.view$
#> 174                                 \\.skyd$
#> 175                                 \\.skyl$
#> 176                                  \\.sne$
#> 177                                \\.htrms$
#> 178                                 \\.cpsx$
#> 179                               \\.PARAMS$
#> 180                               \\.params$
#> 181                                \\.param$
#> 182                                  \\.par$
#> 183                               \\.config$
#> 184                                 \\.apar$
#> 185                                 \\.knwf$
#> 186                                 \\.json$
#> 187                                 \\.toml$
#> 188                                 \\.yaml$
#> 189                                  \\.ini$
#> 190                                  \\.mtd$
#> 191                                \\.index$
#> 192                               \\.method$
#> 193                               \\.Method$
#> 194                             \\.FAmethod$
#> 195                           \\.properties$
#> 196                                    \\.R$
#> 197                                   \\.py$
#> 198                                    \\.r$
#> 199                                   \\.pl$
#> 200                                   \\.js$
#> 201                                  \\.jar$
#> 202                                  \\.Rmd$
#> 203                                   \\.sh$
#> 204                                \\.ipynb$
#> 205                                  \\.exe$
#> 206                                  \\.bin$
#> 207                                  \\.dll$
#> 208                                \\.RData$
#> 209                                  \\.RDS$
#> 210                               \\.sqlite$
#> 211                                   \\.h5$
#> 212                                  \\.md5$
#> 213                                \\.cksum$
#> 214                               \\.chksum$
#> 215                                  \\.bak$
#> 216                             \\.download$
#> 217                           \\.crdownload$
#> 218                           \\.sgdownload$
#> 219                                 \\.temp$
#> 220                                  \\.tmp$
#> 221                                  \\.gtf$
#> 222                                  \\.gff$
#> 223                                \\.fastq$
#> 224                                  \\.vcf$
#> 225                                \\.plink$
#> 226                                  \\.res$

pxFileTypes("foo")
#>   file type
#> 1  foo <NA>
pxFileTypes("foo.mzML")
#>       file type
#> 1 foo.mzML  raw
pxFileTypes("foo.raw")
#>      file   type
#> 1 foo.raw rawbin
pxFileTypes("foo.txt")
#>      file type
#> 1 foo.txt  doc
pxFileTypes("foo.R")
#>    file type
#> 1 foo.R code
pxFileTypes("foo.fasta")
#>        file type
#> 1 foo.fasta  fas

pxFileTypes(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
#>        file type
#> 1       foo <NA>
#> 2  foo.mzML  raw
#> 3     foo.R code
#> 4 foo.fasta  fas