Infer file type

The px_file_types() function inferres mass spectrometry and proteomics file types based on a currated table of file types and associated patterns. This table can be accessed with file_types(). See the examples below for the content of the tables.

The file types table is generated by scripts/make_file_types.R.

file_types()

px_file_types(fls, types = file_types())

Arguments

fls

character() of file names whose types need to be inferred based on their file extenstion.

types

data.frame of file types. Default is file_types().

Value

A data.frame with the filenames and their inferred types.

Author

Laurent Gatto with contributions from Dr. Samuel Wein to identify several file types.

Examples

file_types()
#>             type                                  ext
#> 1        archive                                  zip
#> 2        archive                               tar.gz
#> 3        archive                                  rar
#> 4        archive                                   7z
#> 5        archive                                  tgz
#> 6        archive                                  ZIP
#> 7        archive                                  tar
#> 8        archive                                  tgz
#> 9        archive                                  RAR
#> 10       archive                                  bz2
#> 11       archive                           webarchive
#> 12       archive                                   gz
#> 13       archive                                   xy
#> 14           doc                                  doc
#> 15           doc                                  pdf
#> 16           doc                                  PDF
#> 17           doc                                  ppt
#> 18           doc                                  odt
#> 19           doc                                 docx
#> 20           doc                                 pptx
#> 21           doc                                  rtf
#> 22           doc                                 html
#> 23           doc                                 html
#> 24           doc                                  htm
#> 25           doc                                shtml
#> 26           doc                               readme
#> 27           doc                                  txt
#> 28           doc                                   md
#> 29        binary                                  raw
#> 30        binary                                  Raw
#> 31        binary                                  RAW
#> 32        binary                                    d
#> 33        binary                                d.zip
#> 34        binary                              raw.zip
#> 35        binary                               raw.gz
#> 36        binary                                 wiff
#> 37        binary                                wiff2
#> 38        binary                            wiff.scan
#> 39        binary                         wiff.1.~idx2
#> 40        binary                             wiff.mtd
#> 41        binary                                  t2d
#> 42           raw                                 mzML
#> 43           raw                              mzML.gz
#> 44           raw                             mzML.zip
#> 45           raw                                mzXML
#> 46           raw                             mzXML.gz
#> 47           raw                            mzXML.zip
#> 48           raw                                mzxml
#> 49           raw                                TraML
#> 50           raw                                traML
#> 51           raw                                traml
#> 52           raw                               netCDF
#> 53           raw                               mzData
#> 54           raw                               mzdata
#> 55           raw                                  mz5
#> 56           raw                                imzML
#> 57           pkl                                  mgf
#> 58           pkl                                  MGF
#> 59           pkl                               mgf.gz
#> 60           pkl                               MGF.gz
#> 61           pkl                                  pkl
#> 62           pkl                               pkl.gz
#> 63           pkl                                  PKL
#> 64           fas                                  fas
#> 65           fas                                fasta
#> 66           fas                                   fa
#> 67           fas                                  faa
#> 68           fas                                FASTA
#> 69           fas                                fasts
#> 70        reflib                                 blib
#> 71        reflib                                 elib
#> 72        reflib                                 dlib
#> 73        reflib                                  msp
#> 74            id                            mzIdentML
#> 75            id                                 mzID
#> 76            id                              mzID.gz
#> 77            id                                 mzid
#> 78            id                              mzid.gz
#> 79            id                                  dat
#> 80            id                               dat.gz
#> 81            id                              dat.zip
#> 82            id                                idXML
#> 83            id                                  omx
#> 84            id                                  dta
#> 85            id                              dta.tgz
#> 86            id                          dta.tar.bz2
#> 87           tbl                                  csv
#> 88           tbl                                  tsv
#> 89           tbl                                  xls
#> 90           tbl                                 xlsx
#> 91           tbl                                 XLSX
#> 92           tbl                                 xlsb
#> 93           tbl                               csv.gz
#> 94           tbl                               tsv.gz
#> 95           tbl                               psmtsv
#> 96           tbl                                delim
#> 97           tbl                              tabular
#> 98         mztab                                mztab
#> 99         mztab                             mztab.gz
#> 100        mztab                                mzTab
#> 101        mztab                             mzTab.gz
#> 102        mztab                              mzTabNA
#> 103        mztab                           -mztab.txt
#> 104          fig                                  png
#> 105          fig                                  jpg
#> 106          fig                                 jpeg
#> 107          fig                                 tiff
#> 108          fig                                  TIF
#> 109          fig                                  tif
#> 110          fig                                  gif
#> 111          fig                                  PNG
#> 112          fig                                  JPG
#> 113          fig                                  svg
#> 114          xml                                  xml
#> 115          xml                               xml.gz
#> 116      prophet                               pepXML
#> 117      prophet                              protXML
#> 118       bruker                                  yep
#> 119       bruker                                  baf
#> 120     scaffold                                  sf3
#> 121     scaffold                                 sptm
#> 122     scaffold                                 sfdb
#> 123     scaffold                                 sdia
#> 124     scaffold                                metdb
#> 125           pd                             pdResult
#> 126           pd                                  msf
#> 127           pd                         pdResultView
#> 128           pd                              msfView
#> 129           pd                           pdAnalysis
#> 130           pd                       pdProcessingWF
#> 131           pd                        pdConsensusWF
#> 132           pd                              pdStudy
#> 133           pd                          pdStudy.bak
#> 134      sequest                                  ms1
#> 135      sequest                                  ms2
#> 136      sequest                                  srf
#> 137      sequest                                  out
#> 138      sequest                              out.tgz
#> 139      sequest                          out.tar.bz2
#> 140 proteinpilot                                group
#> 141   progenesis              ProgenesisQIPExperiment
#> 142   progenesis                 ProgenesisQIPArchive
#> 143   progenesis             ProgenesisLcmsExperiment
#> 144   progenesis ProgenesisQIPMultiFractionExperiment
#> 145      skyline                                  sky
#> 146      skyline                             sky.view
#> 147      skyline                                 skyd
#> 148      skyline                                 skyl
#> 149  spectronaut                                  sne
#> 150       params                               params
#> 151       params                                 apar
#> 152       params                                 knwf
#> 153       params                                 json
#> 154         code                                    R
#> 155         code                                   py
#> 156         code                                    r
#> 157         code                                   pl
#> 158         code                                   js
#> 159         code                                  jar
#> 160         code                                  Rmd
#> 161         code                                   sh
#> 162         code                                ipynb
#> 163         data                                RData
#> 164         data                                  RDS
#> 165         data                               sqlite
#> 166         data                                   h5
#> 167          chk                                  md5
#> 168          chk                                cksum
#> 169          chk                               chksum
#> 170          gen                                  gtf
#> 171          gen                                  gff
#> 172          gen                                fastq
#> 173          gen                                  vcf
#> 174          gen                                plink
#> 175          ssv                                  ssv
#> 176          sqt                                  sqt
#> 177          apl                                  apl
#> 178        htrms                                htrms
#> 179         cpsx                                 cpsx
#> 180          res                                  res
#> 181         pcml                                 pcml
#>                                      pattern
#> 1                                    \\.zip$
#> 2                               \\.tar\\.gz$
#> 3                                    \\.rar$
#> 4                                     \\.7z$
#> 5                                    \\.tgz$
#> 6                                    \\.ZIP$
#> 7                                    \\.tar$
#> 8                                    \\.tgz$
#> 9                                    \\.RAR$
#> 10                                   \\.bz2$
#> 11                            \\.webarchive$
#> 12                                    \\.gz$
#> 13                                    \\.xy$
#> 14                                   \\.doc$
#> 15                                   \\.pdf$
#> 16                                   \\.PDF$
#> 17                                   \\.ppt$
#> 18                                   \\.odt$
#> 19                                  \\.docx$
#> 20                                  \\.pptx$
#> 21                                   \\.rtf$
#> 22                                  \\.html$
#> 23                                  \\.html$
#> 24                                   \\.htm$
#> 25                                 \\.shtml$
#> 26                                \\.readme$
#> 27                                   \\.txt$
#> 28                                    \\.md$
#> 29                                   \\.raw$
#> 30                                   \\.Raw$
#> 31                                   \\.RAW$
#> 32                                     \\.d$
#> 33                               \\.d\\.zip$
#> 34                             \\.raw\\.zip$
#> 35                              \\.raw\\.gz$
#> 36                                  \\.wiff$
#> 37                                 \\.wiff2$
#> 38                           \\.wiff\\.scan$
#> 39                      \\.wiff\\.1\\.~idx2$
#> 40                            \\.wiff\\.mtd$
#> 41                                   \\.t2d$
#> 42                                  \\.mzML$
#> 43                             \\.mzML\\.gz$
#> 44                            \\.mzML\\.zip$
#> 45                                 \\.mzXML$
#> 46                            \\.mzXML\\.gz$
#> 47                           \\.mzXML\\.zip$
#> 48                                 \\.mzxml$
#> 49                                 \\.TraML$
#> 50                                 \\.traML$
#> 51                                 \\.traml$
#> 52                                \\.netCDF$
#> 53                                \\.mzData$
#> 54                                \\.mzdata$
#> 55                                   \\.mz5$
#> 56                                 \\.imzML$
#> 57                                   \\.mgf$
#> 58                                   \\.MGF$
#> 59                              \\.mgf\\.gz$
#> 60                              \\.MGF\\.gz$
#> 61                                   \\.pkl$
#> 62                              \\.pkl\\.gz$
#> 63                                   \\.PKL$
#> 64                                   \\.fas$
#> 65                                 \\.fasta$
#> 66                                    \\.fa$
#> 67                                   \\.faa$
#> 68                                 \\.FASTA$
#> 69                                 \\.fasts$
#> 70                                  \\.blib$
#> 71                                  \\.elib$
#> 72                                  \\.dlib$
#> 73                                   \\.msp$
#> 74                             \\.mzIdentML$
#> 75                                  \\.mzID$
#> 76                             \\.mzID\\.gz$
#> 77                                  \\.mzid$
#> 78                             \\.mzid\\.gz$
#> 79                                   \\.dat$
#> 80                              \\.dat\\.gz$
#> 81                             \\.dat\\.zip$
#> 82                                 \\.idXML$
#> 83                                   \\.omx$
#> 84                                   \\.dta$
#> 85                             \\.dta\\.tgz$
#> 86                       \\.dta\\.tar\\.bz2$
#> 87                                   \\.csv$
#> 88                                   \\.tsv$
#> 89                                   \\.xls$
#> 90                                  \\.xlsx$
#> 91                                  \\.XLSX$
#> 92                                  \\.xlsb$
#> 93                              \\.csv\\.gz$
#> 94                              \\.tsv\\.gz$
#> 95                                \\.psmtsv$
#> 96                                 \\.delim$
#> 97                               \\.tabular$
#> 98                                 \\.mztab$
#> 99                            \\.mztab\\.gz$
#> 100                                \\.mzTab$
#> 101                           \\.mzTab\\.gz$
#> 102                              \\.mzTabNA$
#> 103                         \\.-mztab\\.txt$
#> 104                                  \\.png$
#> 105                                  \\.jpg$
#> 106                                 \\.jpeg$
#> 107                                 \\.tiff$
#> 108                                  \\.TIF$
#> 109                                  \\.tif$
#> 110                                  \\.gif$
#> 111                                  \\.PNG$
#> 112                                  \\.JPG$
#> 113                                  \\.svg$
#> 114                                  \\.xml$
#> 115                             \\.xml\\.gz$
#> 116                               \\.pepXML$
#> 117                              \\.protXML$
#> 118                                  \\.yep$
#> 119                                  \\.baf$
#> 120                                  \\.sf3$
#> 121                                 \\.sptm$
#> 122                                 \\.sfdb$
#> 123                                 \\.sdia$
#> 124                                \\.metdb$
#> 125                             \\.pdResult$
#> 126                                  \\.msf$
#> 127                         \\.pdResultView$
#> 128                              \\.msfView$
#> 129                           \\.pdAnalysis$
#> 130                       \\.pdProcessingWF$
#> 131                        \\.pdConsensusWF$
#> 132                              \\.pdStudy$
#> 133                        \\.pdStudy\\.bak$
#> 134                                  \\.ms1$
#> 135                                  \\.ms2$
#> 136                                  \\.srf$
#> 137                                  \\.out$
#> 138                            \\.out\\.tgz$
#> 139                      \\.out\\.tar\\.bz2$
#> 140                                \\.group$
#> 141              \\.ProgenesisQIPExperiment$
#> 142                 \\.ProgenesisQIPArchive$
#> 143             \\.ProgenesisLcmsExperiment$
#> 144 \\.ProgenesisQIPMultiFractionExperiment$
#> 145                                  \\.sky$
#> 146                           \\.sky\\.view$
#> 147                                 \\.skyd$
#> 148                                 \\.skyl$
#> 149                                  \\.sne$
#> 150                               \\.params$
#> 151                                 \\.apar$
#> 152                                 \\.knwf$
#> 153                                 \\.json$
#> 154                                    \\.R$
#> 155                                   \\.py$
#> 156                                    \\.r$
#> 157                                   \\.pl$
#> 158                                   \\.js$
#> 159                                  \\.jar$
#> 160                                  \\.Rmd$
#> 161                                   \\.sh$
#> 162                                \\.ipynb$
#> 163                                \\.RData$
#> 164                                  \\.RDS$
#> 165                               \\.sqlite$
#> 166                                   \\.h5$
#> 167                                  \\.md5$
#> 168                                \\.cksum$
#> 169                               \\.chksum$
#> 170                                  \\.gtf$
#> 171                                  \\.gff$
#> 172                                \\.fastq$
#> 173                                  \\.vcf$
#> 174                                \\.plink$
#> 175                                  \\.ssv$
#> 176                                  \\.sqt$
#> 177                                  \\.apl$
#> 178                                \\.htrms$
#> 179                                 \\.cpsx$
#> 180                                  \\.res$
#> 181                                 \\.pcml$

px_file_types("foo")
#>   file type
#> 1  foo <NA>
px_file_types("foo.mzML")
#>       file type
#> 1 foo.mzML  raw
px_file_types("foo.raw")
#>      file   type
#> 1 foo.raw binary
px_file_types("foo.txt")
#>      file type
#> 1 foo.txt  doc
px_file_types("foo.R")
#>    file type
#> 1 foo.R code
px_file_types("foo.fasta")
#>        file type
#> 1 foo.fasta  fas

px_file_types(c("foo", "foo.mzML", "foo.R", "foo.fasta"))
#>        file type
#> 1       foo <NA>
#> 2  foo.mzML  raw
#> 3     foo.R code
#> 4 foo.fasta  fas